refactor: move macros under macros
